- create branch `migrate-promise-utils` to migrate the promise and task related utility functions inside of `/src/lambda.ts` to a new submodule `/src/promiseman.ts`. - [breaking changes] add a new submodule in `/src/promiseman.ts`, and migrate over the promise and task related utility functions inside of `/src/lambda.ts`. - export the new `/src/promiseman.ts` submodule in `/deno.json`, and `/src/mod.ts`.
