+++ src/mod.ts
@@ -36,6 +36,7 @@ export * from "./mapper.ts"
 export * from "./numericarray.ts"
 export * from "./numericmethods.ts"
 export * from "./pathman.ts"
+export * from "./promiseman.ts"
 // export * from "./semver.ts"
 export * from "./stringman.ts"
 export * from "./struct.ts"
